requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
dnspython==2.1.0
gevent==21.8.0
greenlet==1.1.1
idna==3.2
lxml==4.6.3
rarfile==4.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
subfinder==2.2.1
Sublist3r==1.0
urllib3==1.26.6
zope.event==4.5.0
zope.interface==5.4.0
